Add files via upload
Days Between Dates
PYTHON

The "Days Between Dates" Tkinter application offers a user-friendly interface where users can input two dates to calculate the number of days between them. With a well-designed layout, color scheme, and styling, the application provides a modern and sleek feel. The use of the 'clam' theme for ttk widgets enhances the overall appearance, and the inclusion of a "Calculate" button and clear result display adds to the functionality. Overall, this utility tool is easy to navigate and provides an efficient solution for users to calculate time spans across different operating systems.